LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/uievents/mouse/synthetic-mouse-enter-leave-over-out-button-state-after-target-removed.tentative_buttonType=LEFT&button=0&buttons=1-expected.txt


PASS Setup event listeners and wait for load
FAIL Removing an element at mousedown assert_equals: All events should've been checked expected "[]" but got "[{ 'mousemove' on 'parent': button=0, buttons=0 }]"
FAIL Removing an element at mouseup assert_equals: All events should've been checked expected "[]" but got "[{ 'mousemove' on 'parent': button=0, buttons=0 }]"
PASS Removing an element at mousedown: mousedown should've been fired
FAIL Removing an element at mousedown: mouseout and mouseleave should've been fired on the removed child assert_equals: expected "[{ 'mouseout' on 'child': button=0, buttons=1 }, { 'mouseleave' on 'child': button=0, buttons=1 }]" but got "[]"
FAIL Removing an element at mousedown: mouseover should've been fired on the parent assert_equals: expected "[{ 'mouseover' on 'parent': button=0, buttons=1 }]" but got "[{ 'mouseover' on 'parent': button=0, buttons=0 }]"
FAIL Removing an element at mousedown: mouseenter should not have been fired on the parent assert_equals: expected "[]" but got "[{ 'mouseenter' on 'parent': button=0, buttons=0 }]"
FAIL Removing an element at mouseup: mousedown should've been fired assert_equals: expected "{ 'mousedown' on 'child': button=0, buttons=1 }" but got "{ 'mousedown' on 'parent': button=0, buttons=1 }"
FAIL Removing an element at mouseup: mouseout and mouseleave should've been fired on the removed child assert_equals: expected "[{ 'mouseout' on 'child': button=0, buttons=0 }, { 'mouseleave' on 'child': button=0, buttons=0 }]" but got "[]"
FAIL Removing an element at mouseup: mouseover should've been fired on the parent assert_equals: expected "[{ 'mouseover' on 'parent': button=0, buttons=0 }]" but got "[]"
PASS Removing an element at mouseup: mouseenter should not have been fired on the parent
